research-service/requirements.txt
mysql-connector-python==8.0.33
requests==2.31.0
yfinance==0.2.28
python-dotenv==1.0.0
